requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
opencv-python>=4.5.0
Pillow>=8.0.0
matplotlib>=3.3.0
scikit-learn>=0.24.0
scipy>=1.7.0
torch>=1.10.0
torchvision>=0.11.0

# Optional: Weights & Biases for experiment tracking
# wandb>=0.12.0
